requirements.txt
boto3
google-genai
pillow